Sort of complete UC74 main functions. Now, nutritionist will have to be the designated nutritionist in order to view the patient food diary.
(need to add more test cases associate the new features, and we need the
sequence diagram and those documents.)

Due to not breaking old tests, I just append the error message second
line "Has to be designed nutritionist in order to view patinet's food
diary." instead of changing it.
I still can change it if you guys think we have to change the whole
error message.